Computation of selection envelope is now done on database side

+++ bundles/core-map/src/main/java/org/orbisgis/coremap/layerModel/Layer.java
@@ -256,7 +256,7 @@ public class Layer extends BeanLayer {
 	}
 
 	@Override
-	public void setSelection(Set<Integer> newSelection) {
+	public void setSelection(Set<Long> newSelection) {
 		super.setSelection(newSelection);
 		fireSelectionChanged();
 	}
